requirements.txt
# Core Dependencies
lxml>=4.9.0
pydantic>=2.0.0
jinja2>=3.1.0
click>=8.1.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
slowapi>=0.1.9  # Rate limiting

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Configuration
python-dotenv>=1.0.0

# System Monitoring (for service management)
psutil>=5.9.0

# MQTT Support
paho-mqtt>=1.6.0

# HTTP Client (for startup health checks)
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Required for FastAPI TestClient

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0

# Code Quality
black>=23.0.0
pylint>=2.17.0
mypy>=1.4.0

# Optional - For advanced features
aiofiles>=23.0.0
redis>=4.6.0
celery>=5.3.0
numpy>=1.24.0
matplotlib>=3.7.0

# XML Schema Validation
xmlschema>=2.3.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Error Tracking & Monitoring
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
